Make exoplayer not depend statically on androidx
The androidx annotation library is about to start depending on kotlin,
which, with a static dependency means the kotlin-stdlib is getting
included in the exoplayer jars.

Since the annotations are only used at compile-time, move the dependency
to a classpath dep instead. Longer term, we should not be dependning
on androidx from these jars at all.

Bug: 214255490
Test: m updatable-media (with new androidx)
Change-Id: I51dd549584f4174916ea4be22eb6246e29dff52a