Remove deleted test from TEST_MAPPING.
These tests are being renamed as we moved to Android.bp and this
mapping also needs to be updated. Since the test is a Robolectric test
and these are not support by TEST_MAPPING, simple remove this test.

Note that the existing test was not working anyway, so this is not
really reducing what we are testing.

Test: presubmit
Bug: 330241426
Bug: 340381453
Change-Id: Iccc0af31df1175709b22a47b43684d8a733d24e1